\name{productPowerMinusHalf}
\alias{productPowerMinusHalf}
\title{ 
Square root of product for every elements in a vector
}
\description{
This function computes Square root of product for every elements in a vector.
}
\usage{
productPowerMinusHalf(vec)
}
\arguments{
  \item{vec}{A input vector.}
}

\value{
The square root of product for every elements in a vector
}

\author{
\packageAuthor{SKFCPD}

Maintainer: \packageMaintainer{SKFCPD}
}

\keyword{internal}